traits: Move `SeekStreamLen` to `lofty::util::io`
