+++ src/musepack/read.rs
@@ -6,7 +6,7 @@ use crate::config::ParseOptions;
 use crate::error::Result;
 use crate::id3::v2::read::parse_id3v2;
 use crate::id3::{find_id3v1, find_id3v2, find_lyrics3v2, FindId3v2Config, ID3FindResults};
-use crate::traits::SeekStreamLen;
+use crate::util::io::SeekStreamLen;
 
 use std::io::{Read, Seek, SeekFrom};
