+++ src/util/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod alloc;
+pub(crate) mod io;
 pub(crate) mod math;
 pub(crate) mod text;
